Move DummyDataProvider to relay-lsp
