+++ compiler/crates/relay-lsp/src/lib.rs
@@ -39,7 +39,7 @@ pub use hover::ContentConsumerType;
 pub use js_language_server::JSLanguageServer;
 use log::debug;
 pub use lsp_extra_data_provider::{
-    FieldDefinitionSourceInfo, FieldSchemaInfo, LSPExtraDataProvider,
+    DummyExtraDataProvider, FieldDefinitionSourceInfo, FieldSchemaInfo, LSPExtraDataProvider,
 };
 use lsp_process_error::LSPProcessResult;
 pub use lsp_runtime_error::{LSPRuntimeError, LSPRuntimeResult};
