+++ compiler/crates/relay-lsp/src/lsp_extra_data_provider.rs
@@ -29,3 +29,26 @@ pub trait LSPExtraDataProvider: Send + Sync {
         field_info: Option<FieldSchemaInfo>,
     ) -> Result<Option<FieldDefinitionSourceInfo>, String>;
 }
+
+pub struct DummyExtraDataProvider {}
+
+impl DummyExtraDataProvider {
+    pub fn new() -> DummyExtraDataProvider {
+        DummyExtraDataProvider {}
+    }
+}
+
+impl LSPExtraDataProvider for DummyExtraDataProvider {
+    fn fetch_query_stats(&self, _search_token: &str) -> Vec<String> {
+        vec![]
+    }
+
+    fn resolve_field_definition(
+        &self,
+        _project_name: String,
+        _parent_type: String,
+        _field_info: Option<FieldSchemaInfo>,
+    ) -> Result<Option<FieldDefinitionSourceInfo>, String> {
+        Ok(None)
+    }
+}
